test: split exact domain operation checkers by domain
